# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: taller
    description: Taller state operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User management operations
  - name: roles
    description: Role and permission management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /taller/state:
    get:
      operationId: getTallerState
      tags: [taller]
      summary: Get full taller state
      description: Returns equipos, gpvList, and tecnicos
      responses:
        "200":
          description: Current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TallerState"
    put:
      operationId: saveTallerState
      tags: [taller]
      summary: Save full taller state
      description: Upserts the full state (equipos, gpvList, tecnicos)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TallerStateInput"
      responses:
        "200":
          description: Saved state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TallerState"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Concurrent edit conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conflict"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out the current session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current authenticated user and role
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List all roles
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /roles/{id}:
    put:
      operationId: updateRole
      tags: [roles]
      summary: Update a role
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteRole
      tags: [roles]
      summary: Delete a role
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted
        "400":
          description: Cannot delete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Equipo:
      type: object
      properties:
        id:          { type: number }
        destino:     { type: string }
        modelo:      { type: string }
        interno:     { type: string }
        accesorio:   { type: string }
        cliente:     { type: string }
        fechaIngreso: { type: string }
        falla:       { type: string }
        estado:      { type: string }
        tecnicos:
          type: array
          items: { type: string }
        observacion: { type: string }
        prioridad:   { type: string }
        enVentaDesde:
          type: ["string", "null"]
      required:
        - id
        - destino
        - modelo
        - interno
        - accesorio
        - cliente
        - fechaIngreso
        - falla
        - estado
        - tecnicos
        - observacion
        - prioridad

    GPVEntry:
      type: object
      properties:
        id:           { type: number }
        modelo:       { type: string }
        interno:      { type: string }
        cliente:      { type: string }
        fechaEntrega: { type: string }
        observacion:  { type: string }
        estado:       { type: string }
      required:
        - id
        - modelo
        - interno
        - cliente
        - fechaEntrega
        - observacion
        - estado

    TallerState:
      type: object
      properties:
        equipos:
          type: array
          items:
            $ref: "#/components/schemas/Equipo"
        gpvList:
          type: array
          items:
            $ref: "#/components/schemas/GPVEntry"
        tecnicos:
          type: array
          items: { type: string }
        layout:
          type: object
          additionalProperties:
            type: array
            items: { type: number }
        updatedAt:
          type: ["string", "null"]
      required:
        - equipos
        - gpvList
        - tecnicos
        - layout

    TallerStateInput:
      type: object
      properties:
        equipos:
          type: array
          items:
            $ref: "#/components/schemas/Equipo"
        gpvList:
          type: array
          items:
            $ref: "#/components/schemas/GPVEntry"
        tecnicos:
          type: array
          items: { type: string }
        layout:
          type: object
          additionalProperties:
            type: array
            items: { type: number }
        expectedUpdatedAt:
          type: ["string", "null"]
          description: Last-known updatedAt for optimistic concurrency. If it does not match the stored value, the server returns 409.
      required:
        - equipos
        - gpvList
        - tecnicos
        - layout

    Error:
      type: object
      properties:
        error: { type: string }
      required:
        - error

    Conflict:
      type: object
      properties:
        error: { type: string }
        current:
          $ref: "#/components/schemas/TallerState"
      required:
        - error
        - current

    ModulePermission:
      type: object
      properties:
        view:   { type: boolean }
        create: { type: boolean }
        edit:   { type: boolean }
        delete: { type: boolean }
      required:
        - view
        - create
        - edit
        - delete

    Role:
      type: object
      properties:
        id:   { type: number }
        name: { type: string }
        isSystem: { type: boolean }
        permissions:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ModulePermission"
      required:
        - id
        - name
        - isSystem
        - permissions

    RoleInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        permissions:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ModulePermission"
      required:
        - name
        - permissions

    RoleUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        permissions:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ModulePermission"

    User:
      type: object
      properties:
        id:       { type: number }
        username: { type: string }
        nombre:   { type: string }
        roleId:   { type: number }
        roleName: { type: string }
        activo:   { type: boolean }
      required:
        - id
        - username
        - nombre
        - roleId
        - roleName
        - activo

    UserInput:
      type: object
      properties:
        username: { type: string, minLength: 1 }
        nombre:   { type: string }
        password: { type: string, minLength: 6 }
        roleId:   { type: number }
        activo:   { type: boolean }
      required:
        - username
        - password
        - roleId

    UserUpdate:
      type: object
      properties:
        nombre:   { type: string }
        password: { type: string, minLength: 6 }
        roleId:   { type: number }
        activo:   { type: boolean }

    LoginInput:
      type: object
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
      required:
        - username
        - password

    AuthSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        role:
          $ref: "#/components/schemas/Role"
      required:
        - user
        - role