# lib/api-zod/src/generated/types/conflict.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TallerState } from './tallerState';

export interface Conflict {
  error: string;
  current: TallerState;
}