# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RolePermissions } from './rolePermissions';

export interface Role {
  id: number;
  name: string;
  isSystem: boolean;
  permissions: RolePermissions;
}